# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kokpit Okulları API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /haberler:
    get:
      operationId: listHaberler
      tags: [haberler]
      summary: Tüm haberler
      parameters:
        - name: kategori
          in: query
          required: false
          schema:
            type: string
        - name: yayinda
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Haber listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Haber"
    post:
      operationId: createHaber
      tags: [haberler]
      summary: Yeni haber oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHaberBody"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Haber"
  /haberler/{id}:
    get:
      operationId: getHaber
      tags: [haberler]
      summary: Haber detayı
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Haber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Haber"
        "404":
          description: Bulunamadı
    patch:
      operationId: updateHaber
      tags: [haberler]
      summary: Haberi güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHaberBody"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Haber"
        "404":
          description: Bulunamadı
    delete:
      operationId: deleteHaber
      tags: [haberler]
      summary: Haberi sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
...
  /foto-galeri:
    get:
      operationId: listFotoGaleri
      tags: [fotogaleri]
      summary: Tüm fotoğraflar
      parameters:
        - name: kategori
          in: query
          required: false
          schema:
            type: string
        - name: yayinda
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Fotoğraf listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FotoGaleri"
    post:
      operationId: createFotoGaleri
      tags: [fotogaleri]
      summary: Yeni fotoğraf ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFotoGaleriBody"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FotoGaleri"
  /foto-galeri/{id}:
    patch:
      operationId: updateFotoGaleri
      tags: [fotogaleri]
      summary: Fotoğrafı güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFotoGaleriBody"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FotoGaleri"
        "404":
          description: Bulunamadı
    delete:
      operationId: deleteFotoGaleri
      tags: [fotogaleri]
      summary: Fotoğrafı sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
  /video-galeri:
    get:
      operationId: listVideoGaleri
      tags: [videogaleri]
      summary: Tüm videolar
      parameters:
        - name: kategori
          in: query
          required: false
          schema:
            type: string
        - name: yayinda
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Video listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VideoGaleri"
    post:
      operationId: createVideoGaleri
      tags: [videogaleri]
      summary: Yeni video ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoGaleriBody"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoGaleri"
  /video-galeri/{id}:
    patch:
      operationId: updateVideoGaleri
      tags: [videogaleri]
      summary: Videoyu güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoGaleriBody"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoGaleri"
        "404":
          description: Bulunamadı
    delete:
      operationId: deleteVideoGaleri
      tags: [videogaleri]
      summary: Videoyu sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi
        "404":
          description: Bulunamadı
components:
  schemas:
...
    Haber:
      type: object
      properties:
        id:
          type: integer
        baslik:
          type: string
        ozet:
          type: string
        icerik:
          type: string
        resimUrl:
          type: ["string", "null"]
        kategori:
          type: string
        yayinda:
          type: boolean
        onemli:
          type: boolean
        yayinTarihi:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - baslik
        - ozet
        - icerik
        - kategori
        - yayinda
        - onemli
        - yayinTarihi
        - createdAt
        - updatedAt
    CreateHaberBody:
      type: object
      properties:
        baslik:
          type: string
        ozet:
          type: string
        icerik:
          type: string
        resimUrl:
          type: ["string", "null"]
        kategori:
          type: string
        yayinda:
          type: boolean
        onemli:
          type: boolean
        yayinTarihi:
          type: string
      required:
        - baslik
        - ozet
        - icerik
        - kategori
    UpdateHaberBody:
      type: object
      properties:
        baslik:
          type: string
        ozet:
          type: string
        icerik:
          type: string
        resimUrl:
          type: ["string", "null"]
        kategori:
          type: string
        yayinda:
          type: boolean
        onemli:
          type: boolean
        yayinTarihi:
          type: string
...
    FotoGaleri:
      type: object
      properties:
        id:
          type: integer
        baslik:
          type: string
        resimUrl:
          type: string
        kategori:
          type: string
        yayinda:
          type: boolean
        sira:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - baslik
        - resimUrl
        - kategori
        - yayinda
        - sira
        - createdAt
        - updatedAt
    CreateFotoGaleriBody:
      type: object
      properties:
        baslik:
          type: string
        resimUrl:
          type: string
        kategori:
          type: string
        yayinda:
          type: boolean
        sira:
          type: integer
      required:
        - baslik
        - resimUrl
        - kategori
    UpdateFotoGaleriBody:
      type: object
      properties:
        baslik:
          type: string
        resimUrl:
          type: string
        kategori:
          type: string
        yayinda:
          type: boolean
        sira:
          type: integer
    VideoGaleri:
      type: object
      properties:
        id:
          type: integer
        baslik:
          type: string
        youtubeId:
          type: string
        kategori:
          type: string
        yayinda:
          type: boolean
        sira:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - baslik
        - youtubeId
        - kategori
        - yayinda
        - sira
        - createdAt
        - updatedAt
    CreateVideoGaleriBody:
      type: object
      properties:
        baslik:
          type: string
        youtubeId:
          type: string
        kategori:
          type: string
        yayinda:
          type: boolean
        sira:
          type: integer
      required:
        - baslik
        - youtubeId
        - kategori
    UpdateVideoGaleriBody:
      type: object
      properties:
        baslik:
          type: string
        youtubeId:
          type: string
        kategori:
          type: string
        yayinda:
          type: boolean
        sira:
          type: integer