# lib/db/src/schema/haberler.ts
import { pgTable, text, serial, timestamp, boolean, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const haberlerTable = pgTable("haberler", {
  id: serial("id").primaryKey(),
  baslik: text("baslik").notNull(),
  ozet: text("ozet").notNull(),
  icerik: text("icerik").notNull(),
  resimUrl: text("resim_url"),
  kategori: text("kategori").notNull().default("genel"),
  yayinda: boolean("yayinda").notNull().default(true),
  onemli: boolean("onemli").notNull().default(false),
  yayinTarihi: timestamp("yayin_tarihi", { withTimezone: true }).notNull().defaultNow(),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertHaberSchema = createInsertSchema(haberlerTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertHaber = z.infer<typeof insertHaberSchema>;
export type Haber = typeof haberlerTable.$inferSelect;